misc: added tests
