+++ crates/mun_project/src/manifest.rs
@@ -28,7 +28,8 @@ impl Manifest {
     /// Try to read a manifest from a file
     pub fn from_file<P: AsRef<Path>>(path: P) -> Result<Manifest, anyhow::Error> {
         // Load the contents of the file
-        let file_contents = std::fs::read_to_string(path)?;
+        let file_contents = std::fs::read_to_string(path)
+            .map_err(|e| anyhow::anyhow!("could not read manifest file: {}", e))?;
         Self::from_str(&file_contents)
     }
 
@@ -39,12 +40,12 @@ impl Manifest {
 
     /// Returns the name of the package
     pub fn name(&self) -> &str {
-        &self.package_id.name
+        &self.package_id.name()
     }
 
     /// Returns the version of the package
     pub fn version(&self) -> &semver::Version {
-        &self.package_id.version
+        &self.package_id.version()
     }
 
     /// Returns the metadata information of the package
@@ -67,7 +68,7 @@ impl PackageId {
 
 impl fmt::Display for PackageId {
     fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        write!(f, "{} v{}", self.name, self.version)
+        write!(f, "{} v{}", self.name(), self.version())
     }
 }
 
@@ -76,7 +77,8 @@ impl std::str::FromStr for Manifest {
 
     fn from_str(s: &str) -> Result<Self, Self::Err> {
         // Parse the contents of the file to toml manifest
-        let manifest = ::toml::from_str::<toml::TomlManifest>(s)?;
+        let manifest = ::toml::from_str::<toml::TomlManifest>(s)
+            .map_err(|e| anyhow::anyhow!("could not parse manifest: {}", e))?;
         manifest.into_real_manifest()
     }
 }
@@ -103,7 +105,7 @@ mod tests {
             manifest.version(),
             &semver::Version::from_str("0.2.0").unwrap()
         );
-        assert_eq!(manifest.metadata.authors, vec!["Mun Team"]);
+        assert_eq!(manifest.metadata().authors, vec!["Mun Team"]);
         assert_eq!(format!("{}", manifest.package_id()), "test v0.2.0");
     }
 }
