+++ crates/mun_project/tests/resources/mun.toml
@@ -1,0 +1,4 @@
+[package]
+name="test"
+version="0.2.0"
+authors = ["Mun Team"]
