+++ crates/mun_project/tests/resources/src/main.mun
@@ -1,0 +1,2 @@
+pub fn main() {
+}
